#1363 Move PoolType to sharding-core.
